Updated __init__.py in subpackages
Updated so that import statements in messages could be cleaner, using the name of the package and the name of the imported class.

Also, refractored the messages_prompts.py as prompts.py, and included a Prompts class so that the origin of the prompts static imports was more clear in message.py.